feat: fair TreeSHAP ensemble comparison that propagates to deployment
Fork of DUNE that makes the Stage-1 model choice reach the deployed
sub-models, and compares four tree ensembles fairly for in-network
volumetric DoS detection on TON-IoT.

- swap PCFI for TreeSHAP: model-agnostic per-class importance; compare
  Random Forest, XGBoost, LightGBM, CatBoost as the Stage-1 model
- normalize per-class TreeSHAP to sum 1 (PCFI contract the SPP gain assumes)
- Stage 4 ranks features by the cluster TreeSHAP over BMv2-deployable
  features instead of a fresh RF Gini, so the Stage-1 model propagates to
  the deployed features, tree configs and sequence
- FAIR_FEATURES flag (env DUNE_ARENA_FAIR_FEATURES=0) toggles the fix off to
  reproduce DUNE original behaviour, where models with the same partition
  deploy identically (the models-behave-equally bug this work documents)
- flow-grouped Stage-1 train/val split: no within-flow leakage
- brute-force TSP replacing the gurobi dependency; n-class voting-table
  generator; py3.12 / pandas2 / numpy2 compatibility
- treeshap/: end-to-end pipeline glue + single paths.py (env override) +
  corrected results summary (collate)
- testbed/: vendored BMv2/Mininet testbed from nds-group/DUNE-bmv2
- uv packaging (pyproject.toml + uv.lock); drop dune-tofino, UNSW, notebooks

+++ model_sequencing/src/model_sequencing.py
@@ -150,7 +150,7 @@ def analyze_model(use_case, classes_filter, npkts, n_tree, max_leaf, feats, clas
     test_flow_IDs = test_data['Flow ID'].to_list()
 
     train_samples = train_data[feats]
-    train_labels = train_data['Label_NEW'].replace(classes, range(len(classes)))
+    train_labels = train_data['Label_NEW'].replace(classes, range(len(classes))).astype(int)
 
     test_samples = test_data[feats]
     test_labels = test_data['Label_NEW'].replace(classes, range(len(classes)))
